+++ packages/astro/test/fixtures/csp/astro.config.mjs
@@ -1,0 +1,10 @@
+// @ts-check
+import { defineConfig } from 'astro/config';
+import react from '@astrojs/react';
+
+export default defineConfig({
+	experimental: {
+		csp: true,
+	},
+	integrations: [react()],
+});