+++ packages/astro/test/fixtures/custom-500-middleware/src/middleware.js
@@ -1,0 +1,4 @@
+export function onRequest(_context, next) {
+    throw 'an error'
+    return next()
+}