+++ packages/integrations/sitemap/src/write-sitemap.ts
@@ -1,24 +1,24 @@
-import { type WriteStream, createWriteStream } from 'node:fs';
+import { createWriteStream, type WriteStream } from 'node:fs';
 import { mkdir } from 'node:fs/promises';
 import { normalize, resolve } from 'node:path';
-import { Readable, pipeline } from 'node:stream';
+import { pipeline, Readable } from 'node:stream';
 import { promisify } from 'node:util';
-import replace from 'stream-replace-string';
-
-import { SitemapAndIndexStream, SitemapStream } from 'sitemap';
-
 import type { AstroConfig } from 'astro';
+import { SitemapAndIndexStream, SitemapIndexStream, SitemapStream } from 'sitemap';
+import replace from 'stream-replace-string';
 import type { SitemapItem } from './index.js';
 
 type WriteSitemapConfig = {
 	filenameBase: string;
 	hostname: string;
 	sitemapHostname?: string;
+	customSitemaps?: string[];
 	sourceData: SitemapItem[];
 	destinationDir: string;
 	publicBasePath?: string;
 	limit?: number;
 	xslURL?: string;
+	lastmod?: string;
 };
 
 // adapted from sitemap.js/sitemap-simple
@@ -30,8 +30,10 @@ export async function writeSitemap(
 		sourceData,
 		destinationDir,
 		limit = 50000,
+		customSitemaps = [],
 		publicBasePath = './',
 		xslURL: xslUrl,
+		lastmod,
 	}: WriteSitemapConfig,
 	astroConfig: AstroConfig,
 ) {
@@ -65,11 +67,20 @@ export async function writeSitemap(
 				stream = sitemapStream.pipe(createWriteStream(writePath));
 			}
 
-			return [new URL(publicPath, sitemapHostname).toString(), sitemapStream, stream];
+			const url = new URL(publicPath, sitemapHostname).toString();
+			return [{ url, lastmod }, sitemapStream, stream];
 		},
 	});
 
 	const src = Readable.from(sourceData);
 	const indexPath = resolve(destinationDir, `./${filenameBase}-index.xml`);
+	for (const url of customSitemaps) {
+		SitemapIndexStream.prototype._transform.call(
+			sitemapAndIndexStream,
+			{ url, lastmod },
+			'utf8',
+			() => {},
+		);
+	}
 	return promisify(pipeline)(src, sitemapAndIndexStream, createWriteStream(indexPath));
 }