+++ packages/integrations/web-vitals/src/constants.ts
@@ -1,1 +1,0 @@
-export const WEB_VITALS_ENDPOINT_PATH = '/_web-vitals';